requirements.txt
streamlit
numpy
opencv-python
pillow
tensorflow
matplotlib
